backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
PyMySQL==1.1.1
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.9.0
python-multipart==0.0.9
